// apps/fcm/gen/mcs.ts
export interface HeartbeatStat {
  ip?: string | undefined;
  timeout?: boolean | undefined;
  intervalMs?: number | undefined;
}

export interface HeartbeatConfig {
  uploadStat?: boolean | undefined;
  ip?: string | undefined;
  intervalMs?: number | undefined;
}

export interface ErrorInfo {
  code?: number | undefined;
  message?: string | undefined;
  type?: string | undefined;
  extension?: Extension | undefined;
}

export interface Setting {
  name?: string | undefined;
  value?: string | undefined;
}

export interface Extension {
  id?: number | undefined;
  data?: Uint8Array | undefined;
}

export interface ClientEvent {
  type?: ClientEvent_Type | undefined;
  numberDiscardedEvents?: number | undefined;
  networkType?: number | undefined;
  timeConnectionStartedMs?: string | undefined;
  timeConnectionEndedMs?: string | undefined;
  errorCode?: number | undefined;
  timeConnectionEstablishedMs?: string | undefined;
}

export enum ClientEvent_Type {
  UNKNOWN = 0,
  DISCARDED_EVENTS = 1,
  FAILED_CONNECTION = 2,
  SUCCESSFUL_CONNECTION = 3,
  UNRECOGNIZED = -1,
}
// ...
export interface LoginRequest {
  id?: string | undefined;
  domain?: string | undefined;
  user?: string | undefined;
  resource?: string | undefined;
  authToken?: string | undefined;
  deviceId?: string | undefined;
  lastRmqId?: string | undefined;
  setting?: Setting[] | undefined;
  compress?: number | undefined;
  receivedPersistentId?: string[] | undefined;
  adaptiveHeartbeat?: boolean | undefined;
  heartbeatStat?: HeartbeatStat | undefined;
  useRmq2?: boolean | undefined;
  accountId?: string | undefined;
  authService?: LoginRequest_AuthService | undefined;
  networkType?: number | undefined;
  status?: string | undefined;
  clientEvent?: ClientEvent[] | undefined;
}

export enum LoginRequest_AuthService {
  ANDROID_ID = 2,
  UNRECOGNIZED = -1,
}
// ...
export interface IqStanza {
  rmqId?: string | undefined;
  type?: IqStanza_IqType | undefined;
  id?: string | undefined;
  from?: string | undefined;
  to?: string | undefined;
  error?: ErrorInfo | undefined;
  extension?: Extension | undefined;
  persistentId?: string | undefined;
  streamId?: number | undefined;
  lastStreamIdReceived?: number | undefined;
  accountId?: string | undefined;
  status?: string | undefined;
}

export enum IqStanza_IqType {
  GET = 0,
  SET = 1,
  RESULT = 2,
  ERROR = 3,
  UNRECOGNIZED = -1,
}
// ...
export interface SelectiveAck {
  id?: string[] | undefined;
}
// ...
export const ErrorInfo: MessageFns<ErrorInfo> = {
  encode(message: ErrorInfo, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.code !== undefined) {
      writer.uint32(8).int32(message.code);
    }
    if (message.message !== undefined) {
      writer.uint32(18).string(message.message);
    }
    if (message.type !== undefined) {
      writer.uint32(26).string(message.type);
    }
    if (message.extension !== undefined) {
      Extension.encode(message.extension, writer.uint32(34).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ErrorInfo {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseErrorInfo();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.code = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.message = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.type = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.extension = Extension.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ErrorInfo {
    return {
      code: isSet(object.code) ? globalThis.Number(object.code) : 0,
      message: isSet(object.message) ? globalThis.String(object.message) : "",
      type: isSet(object.type) ? globalThis.String(object.type) : "",
      extension: isSet(object.extension) ? Extension.fromJSON(object.extension) : undefined,
    };
  },

  toJSON(message: ErrorInfo): unknown {
    const obj: any = {};
    if (message.code !== undefined && message.code !== 0) {
      obj.code = Math.round(message.code);
    }
    if (message.message !== undefined && message.message !== "") {
      obj.message = message.message;
    }
    if (message.type !== undefined && message.type !== "") {
      obj.type = message.type;
    }
    if (message.extension !== undefined) {
      obj.extension = Extension.toJSON(message.extension);
    }
    return obj;
  },

  create(base?: DeepPartial<ErrorInfo>): ErrorInfo {
    return ErrorInfo.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<ErrorInfo>): ErrorInfo {
    const message = createBaseErrorInfo();
    message.code = object.code ?? 0;
    message.message = object.message ?? "";
    message.type = object.type ?? "";
    message.extension = (object.extension !== undefined && object.extension !== null)
      ? Extension.fromPartial(object.extension)
      : undefined;
    return message;
  },
};

function createBaseSetting(): Setting {
  return { name: "", value: "" };
}

export const Setting: MessageFns<Setting> = {
  encode(message: Setting, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.name !== undefined) {
      writer.uint32(10).string(message.name);
    }
    if (message.value !== undefined) {
      writer.uint32(18).string(message.value);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Setting {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSetting();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.name = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.value = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Setting {
    return {
      name: isSet(object.name) ? globalThis.String(object.name) : "",
      value: isSet(object.value) ? globalThis.String(object.value) : "",
    };
  },

  toJSON(message: Setting): unknown {
    const obj: any = {};
    if (message.name !== undefined && message.name !== "") {
      obj.name = message.name;
    }
    if (message.value !== undefined && message.value !== "") {
      obj.value = message.value;
    }
    return obj;
  },

  create(base?: DeepPartial<Setting>): Setting {
    return Setting.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<Setting>): Setting {
    const message = createBaseSetting();
    message.name = object.name ?? "";
    message.value = object.value ?? "";
    return message;
  },
};

function createBaseExtension(): Extension {
  return { id: 0, data: new Uint8Array(0) };
}

export const Extension: MessageFns<Extension> = {
  encode(message: Extension, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== undefined) {
      writer.uint32(8).int32(message.id);
    }
    if (message.data !== undefined && message.data.length !== 0) {
      writer.uint32(18).bytes(message.data);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Extension {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseExtension();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.id = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.data = reader.bytes();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Extension {
    return {
      id: isSet(object.id) ? globalThis.Number(object.id) : 0,
      data: isSet(object.data) ? bytesFromBase64(object.data) : new Uint8Array(0),
    };
  },

  toJSON(message: Extension): unknown {
    const obj: any = {};
    if (message.id !== undefined && message.id !== 0) {
      obj.id = Math.round(message.id);
    }
    if (message.data !== undefined && message.data.length !== 0) {
      obj.data = base64FromBytes(message.data);
    }
    return obj;
  },

  create(base?: DeepPartial<Extension>): Extension {
    return Extension.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<Extension>): Extension {
    const message = createBaseExtension();
    message.id = object.id ?? 0;
    message.data = object.data ?? new Uint8Array(0);
    return message;
  },
};
// ...
export const SelectiveAck: MessageFns<SelectiveAck> = {
  encode(message: SelectiveAck, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== undefined && message.id.length !== 0) {
      for (const v of message.id) {
        writer.uint32(10).string(v!);
      }
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SelectiveAck {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSelectiveAck();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          const el = reader.string();
          if (el !== undefined) {
            message.id!.push(el);
          }
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SelectiveAck {
    return { id: globalThis.Array.isArray(object?.id) ? object.id.map((e: any) => globalThis.String(e)) : [] };
  },

  toJSON(message: SelectiveAck): unknown {
    const obj: any = {};
    if (message.id?.length) {
      obj.id = message.id;
    }
    return obj;
  },

  create(base?: DeepPartial<SelectiveAck>): SelectiveAck {
    return SelectiveAck.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<SelectiveAck>): SelectiveAck {
    const message = createBaseSelectiveAck();
    message.id = object.id?.map((e) => e) || [];
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}